+++ frontend/src/hooks/review/index.ts
@@ -1,3 +1,2 @@
-export * from './writingCardForm';
 export { default as useGetDetailedReview } from './useGetDetailedReview';
 export { default as useGetReviewList } from './useGetReviewList';
