--- frontend/src/pages/ReviewWritingPage/components/index.tsx
@@ -1,8 +1,0 @@
-export { default as AnswerListRecheckModal } from './AnswerListRecheckModal';
-export { default as CardForm } from './CardForm';
-export { default as CardSliderController } from './CardSliderController';
-export { default as ProgressBar } from './ProgressBar';
-export { default as QnABox } from './QnABox';
-export { default as ReviewWritingCard } from './ReviewWritingCard';
-export { default as SubmitCheckModal } from './SubmitCheckModal';
-export { default as NavigateBlockerModal } from './NavigateBlockerModal';
