+++ frontend/src/pages/ReviewWritingPage/constants/modal.ts
@@ -1,0 +1,5 @@
+export const CARD_FORM_MODAL_KEY = {
+  submitConfirm: 'SUBMIT_CONFIRM',
+  navigateConfirm: 'NAVIGATE_CONFIRM',
+  recheck: 'RECHECK',
+};
