+++ frontend/src/pages/ReviewWritingPage/form/components/index.tsx
@@ -1,0 +1,5 @@
+export { default as CardForm } from './CardForm';
+export { default as MultipleChoiceAnswer } from './MultipleChoiceAnswer';
+export { default as QnABox } from './QnABox';
+export { default as ReviewWritingCard } from './ReviewWritingCard';
+export { default as TextAnswer } from './TextAnswer';
