+++ frontend/src/pages/ReviewWritingPage/form/hooks/answers/index.ts
@@ -1,0 +1,4 @@
+export { default as useMultipleChoice } from './multipleChoice/useMultipleChoice';
+export { default as useTextAnswer } from './useTextAnswer';
+export { default as useUpdateDefaultAnswers } from './useUpdateDefaultAnswers';
+export { default as useUpdateReviewerAnswer } from './useUpdateReviewerAnswer';
