+++ frontend/src/pages/ReviewWritingPage/modals/components/AnswerListRecheckModal/components/ReviewCard/styles.ts
