+++ frontend/src/pages/ReviewWritingPage/modals/components/NavigateBlockerModal/style.ts
