+++ frontend/src/pages/ReviewWritingPage/modals/components/SubmitCheckModal/style.ts
