+++ frontend/src/pages/ReviewWritingPage/modals/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as AnswerListRecheckModal } from './AnswerListRecheckModal';
+export { default as CardFormModalContainer } from './CardFormModalContainer';
 export { default as NavigateBlockerModal } from './NavigateBlockerModal';
 export { default as SubmitCheckModal } from './SubmitCheckModal';
