+++ frontend/src/pages/ReviewWritingPage/progressBar/components/ProgressBar/styles.ts
