+++ frontend/src/pages/ReviewWritingPage/slider/components/CardSlider/style.ts
