+++ frontend/src/pages/ReviewWritingPage/slider/components/CardSliderController/index.tsx
