+++ frontend/src/pages/ReviewWritingPage/slider/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as CardSlider } from './CardSlider';
+export { default as CardSliderController } from './CardSliderController';
