+++ frontend/src/pages/ReviewWritingPage/slider/hooks/index.ts
