+++ frontend/src/pages/ReviewWritingPage/slider/hooks/useMovingStepAvailability.ts
