+++ frontend/src/pages/ReviewWritingPage/slider/hooks/useSlideWidthAndHeight.ts
