+++ frontend/src/pages/ReviewWritingPage/types/index.ts
@@ -1,0 +1,1 @@
+export type Direction = 'prev' | 'next' | number;
